src/requirements.txt
openpyxl==3.0.10
python-dotenv==0.20.0
pyyaml==6.0
plaid-python
pymongo